refactor(ui): start api.ts split — move shared types to lib/api/types.ts
First slice of the 662-LOC api.ts decomposition. Hoist the exported
API response interfaces (PaginatedResponse, CoordinatorInviteResult,
ImportInviteResult, AcceptDiscoveredPeerResult, SyncRunItem,
SyncRunResponse, RuntimeInfo, PackTraceCandidate, PackTrace) into a
dedicated types module. api.ts re-exports them so existing
`import type { ... } from "../lib/api"` call sites stay on the same
path, and `import * as api from` keeps access to the response types
via the barrel.
